factor convert social media function into functions folder
